ml-core/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "aether-ml-core"
version = "0.1.0"
description = "ML core for aerial recognition and satellite change detection."
authors = [{ name = "Aether Eye", email = "dev@example.com" }]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "torch",
    "torchvision",
    "timm>=0.9.0",
    "numpy>=1.26.0",
    "PyYAML>=6.0.1",
    "opencv-python-headless>=4.9.0.80",
    "pydantic>=2.5.0",
    "Pillow>=10.0.0",
    "tqdm>=4.66.0",
    "ultralytics>=8.1.0",
    "onnxruntime>=1.17.0",
    "grad-cam>=1.5.5",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
]

[project.urls]
Homepage = "https://example.com"

